Allow popover groups to be controlled from anywhere
