+++ src/components/PopOver/PPopOver.vue
@@ -19,8 +19,8 @@
 
 <script lang="ts" setup>
   import { ref, watch, computed, useAttrs, onMounted, onUnmounted } from 'vue'
-  import { usePopOverGroup } from '@/components/PopOver/compositions/usePopOverGroup'
   import { useMostVisiblePositionStyles } from '@/compositions/position'
+  import { usePopOverGroup } from '@/compositions/usePopOverGroup'
   import { PositionMethod } from '@/types/position'
   import { left, right, bottom, top } from '@/utilities/position'
 
@@ -42,7 +42,7 @@
   const visible = ref(false)
   const attrs = useAttrs()
   const container = ref<Element>()
-  const { setGroupCloseMethod } = usePopOverGroup()
+  const { open: setGroupCloseMethod } = usePopOverGroup(props.group)
 
   const placements = computed(() => Array.isArray(props.placement) ? props.placement : [props.placement])
   const { target, content, styles } = useMostVisiblePositionStyles(placements, { container })
@@ -85,7 +85,7 @@
   }
 
   function open(): void {
-    setGroupCloseMethod(props.group, close)
+    setGroupCloseMethod(close)
 
     container.value = getContainer()
     visible.value = true
