+++ src/compositions/index.ts
@@ -4,5 +4,6 @@ export * from './useColorTheme'
 export * from './useDrawer'
 export * from './useGlobalEventListener'
 export * from './useModal'
+export * from './usePopOverGroup'
 export * from './useScrollLinking'
 export * from './wizard'
